requirements.txt
torch==2.0.1
transformers==4.33.0
scikit-learn==1.3.0
openprompt==1.0.1
PyYAML==6.0.1
numpy==1.25.2
sentencepiece==0.1.96